requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.1
imgaug>=0.4.0
pillow>=10.0.0

# Object Detection
ultralytics>=8.0.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
pycocotools>=2.0.7

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.5.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Data Loading
datasets>=2.14.0
huggingface-hub>=0.16.0

# Metrics
torchmetrics>=1.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
pylint>=2.17.0
isort>=5.12.0

# Type Hints
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Performance
timm>=0.9.0
einops>=0.6.0

# Memory profiling
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional utilities
colorama>=0.4.6
rich>=13.5.0
